Sentinel Application/app/backend/requirements.txt
fastapi==0.115.0
python-multipart==0.0.9
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
anthropic==0.34.0
groq==1.0.0
Pillow==10.4.0
mutagen==1.47.0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
